+++ packages/stubee2-rolling-design-token/src/RollingPalette.ts
@@ -1,0 +1,49 @@
+const text = {
+  Ligther: "#72777A",
+  Light: "#6C7072",
+  Base: "#404446",
+  Dark: "#303437",
+  Darker: "#202325",
+  Darkest: "#090A0A",
+};
+
+const unEmphasize = {
+  Lightest: "#F7F9FA",
+  Lighter: "#F2F4F5",
+  Light: "#DEE0E6",
+  Base: "#BDC2D0",
+  Dark: "#737B98",
+};
+
+const main = {
+  Lightest: "#D5E4F2",
+  Lighter: "#99B3EF",
+  Light: "#5667C1",
+  Base: "#4869F6",
+  Dark: "#1C1F5B",
+  Darkest: "#080C2B",
+};
+
+const error = {
+  Lightest: "#FFC7C4",
+  Lighter: "#FF918B",
+  Light: "#FF4E43",
+  Base: "#F1281C",
+  Dark: "#D61509",
+};
+
+const check = {
+  Lightest: "#ECFCE5",
+  Lighter: "#7DDE86",
+  Light: "#4CD471",
+  Base: "#17C666",
+  Dark: "#198155",
+};
+
+export const RollingPalette = {
+  text,
+  unEmphasize,
+  main,
+  error,
+  check,
+};