+++ packages/stubee2-rolling-ui/src/components/Input/style.ts
@@ -1,7 +1,0 @@
-import styled, { CSSObject } from "styled-components";
-
-export const InputContainer = styled.input<{
-  customStyle?: CSSObject;
-}>`
-  ${({ customStyle }) => customStyle}
-`;