+++ packages/stubee2-rolling-ui/src/components/Ui/Button/Button.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+import { ButtonProps } from "./type";
+import * as S from "./style";
+
+export const Button = ({ ...attr }: ButtonProps) => {
+  return <S.Button {...attr} />;
+};