+++ packages/stubee2-rolling-ui/src/components/Ui/Button/style.ts
@@ -15,22 +15,6 @@ export const Button = styled.button<{
 `;
 
 const getButton: Record<ButtonType, FlattenSimpleInterpolation> = {
-  github: css`
-    width: 515px;
-    height: 75px;
-    outline: none;
-    border: none;
-    border-radius: 15px;
-    background-color: #000;
-    color: #fff;
-    font-size: 20px;
-    font-weight: bold;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    column-gap: 10px;
-    padding: 1rem 2rem;
-  `,
   review: css`
     width: 153px;
     height: 52px;
@@ -43,6 +27,7 @@ const getButton: Record<ButtonType, FlattenSimpleInterpolation> = {
     font-weight: 400;
     box-shadow: 1px 1px 2px rgba(0, 0, 0, 0.5);
   `,
+  custom: css``,
   default: css`
     width: 130px;
     height: 55px;