+++ packages/stubee2-rolling-ui/src/components/Ui/DropDownSelect/type.ts
@@ -1,0 +1,9 @@
+import { Dispatch, SetStateAction } from "react";
+
+export interface DrowDownSelectProps {
+  dropDownSelectItems: string[];
+  width: string;
+  height: string;
+  selectItem: string;
+  setSelectItem: Dispatch<SetStateAction<string>>;
+}