+++ packages/stubee2-rolling-ui/src/index.ts
@@ -1,2 +1,2 @@
-export * from "./components/Button/Button";
-export * from "./components/Input/Input";
+export * from "./components/Layout";
+export * from "./components/Ui";