+++ packages/stubee2-rolling-util/src/hooks/useAuthTopScroll.ts
@@ -2,8 +2,6 @@ import { useEffect } from "react";
 
 export const useAuthTopScroll = () => {
   useEffect(() => {
-    window.onbeforeunload = () => {
-      window.scrollTo(0, 0);
-    };
+    window.scrollTo(0, 0);
   }, []);
 };