+++ packages/stubee2-rolling-util/src/hooks/useCloseModal.ts
@@ -1,9 +1,19 @@
 import { Dispatch, SetStateAction, useEffect, KeyboardEvent } from "react";
 
-export const useCloseModal = (setState: Dispatch<SetStateAction<boolean>>) => {
+export const useCloseModal = (
+  setState: Dispatch<SetStateAction<boolean>>,
+  question?: string
+) => {
   const handleKeyDown = (e: KeyboardEvent<Element>) => {
     if (e.key === "Escape") {
-      setState(false);
+      if (question) {
+        const answer = window.confirm(question);
+        if (answer) {
+          setState(false);
+        }
+      } else {
+        setState(false);
+      }
     }
   };
 
@@ -12,6 +22,7 @@ export const useCloseModal = (setState: Dispatch<SetStateAction<boolean>>) => {
       handleKeyDown(e as unknown as KeyboardEvent<Element>);
     window.addEventListener("keydown", eventListener);
     return () => {
+      document.body.style.overflow = "unset";
       window.removeEventListener("keydown", eventListener);
     };
   }, []);