+++ packages/stubee2-rolling-util/CHANGELOG.md
@@ -1,5 +1,78 @@
 # @stubee2/stubee2-rolling-util
 
+## 1.2.2
+
+### Patch Changes
+
+- fix getTimeAgo param type
+
+## 1.2.1
+
+### Patch Changes
+
+- release
+
+## 1.2.0
+
+### Minor Changes
+
+- releasse
+
+## 1.1.1
+
+### Patch Changes
+
+- release
+- release
+
+## 1.1.0
+
+### Minor Changes
+
+- release
+
+## 1.0.3
+
+### Patch Changes
+
+- fix bug
+
+## 1.0.2
+
+### Patch Changes
+
+- fix bug
+
+## 1.0.1
+
+### Patch Changes
+
+- fix modal
+
+## 1.0.0
+
+### Major Changes
+
+- release
+
+## 0.2.2
+
+### Patch Changes
+
+- release
+
+## 0.2.1
+
+### Patch Changes
+
+- add modalState
+
+## 0.2.0
+
+### Minor Changes
+
+- apply pnp
+
 ## 0.1.0
 
 ### Minor Changes